+++ projects/hono-mcp-server/src/features/base64Encoding.ts
