+++ projects/hono-mcp-server/src/features/getCurrentTime.ts
